backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.12.0
PyPDF2==3.0.1
python-dotenv==1.0.0
pydantic>=2.0.0
httpx>=0.24.0